Fixing a bug in typespec.
The type spec for the init options was specified in such a way that it
was invalid (according to the spec) to pass anything but
`Cassette.Support` as the value for `cassette` and anything but
`AuthenticationHandler` as the value for `handler`. As a result, and
project using Cassette.Plug and overriding those defaults will have
dialyzer errors. Instead of hard-coding the values in the spec, it
should just specify that those values should be module names.